requirements.txt
beautifulsoup4==4.4.1
html5lib==0.999
httplib2==0.9.1
PySocks==1.6.8
requests==2.9.1
SQLAlchemy==1.2.8
cfscrape
urllib3==1.13.1

